app/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0

# MLOps and monitoring
# langfuse==3.0.8  # REMOVED - Using local monitoring instead
prometheus-client==0.19.0

# Core utilities
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
pandas==2.1.4
numpy==1.24.4
scikit-learn==1.3.2

# Dashboard and templating
jinja2==3.1.2
aiofiles

# Background tasks
schedule==1.2.0

# MQTT support
asyncio-mqtt==0.16.1

# PostgreSQL support
psycopg2-binary==2.9.9

# OpenAI API
openai==1.3.8
httpx==0.24.1

# Redis support
redis==5.0.1
aioredis==2.0.1

# SQLAlchemy
sqlalchemy==2.0.23

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
